entitlement: send future events
Analytics needs to know when future changes are requested.
Hook into the AuditedEntitlementDao and send an event on the bus right
after adding the future change on the notification queue.

Although it seems to work, there are two problems with this patch.
First, it adds some more logic in the (already complicated) dao.
Second, I am jumping through hoops to generate the actual event: since I
am in the middle of the transaction, I need to add some logic to find
the previous state for a given event the subscription is in (some of
that state might be on disk, some might still be in memory, in the transaction).

I am open to suggestions to refactor this.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>