invoice: change attributes from DateTime to LocalDate
Make invoice and invoice items dates, in the account timezone,
since the granularity of invoice is the day.

Change also fixed price items to not have an end day for the
service period. This modified the CTD - before, it was set to 30
days in the future for a trial, now it is set to the day the trial
starts.

Notes:
* the BCD is still computed in UTC
* the CTD is rounded to a day

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>