util: add container objects for audit logs
As we start exposing audit logs in the JAX-RS APIs, it becomes clear
that we need some sort of objects to manipulates these logs.

This patch introduces the first version of these objects. No doubt the
interfaces will evolve as we use them more and more. This first pass
simply maps object ids to a list of AuditLog objects.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>