payment: post an event on the bus if there is no default payment method
If a payment cannot be processed because there is no default payment
method, we need to post an event on the bus for overdue to refresh the
overdue state.

Note that an event will be posted even if the invoice balance is zero.

Added a Beatrix test for this scenario.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>