analytics: remove table level comments
MySQL and H2 have a different syntax for table level comments.

MySQL supports creating and updating the comment as part of the CREATE or ALTER
statement.
H2 only supports the syntax COMMENT ON TABLE ... IS '...';

It seems overkill to fork the ddl file just for the comments and these
tables are documented at http://killbilling.org/ anyways.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>