osgi: stop adding lots of bundles in the System bundle
This breaks the OSGI model as they all run in the System bundle:
bundles can't access their original resources for example.

Instead, rely on Felix fileinstall bundle:

  http://felix.apache.org/documentation/subprojects/apache-felix-file-install.html

Unsure if this is the right long-term approach, but it gives us a decent
deployment story for now.

I tried the obr approach as well, but it didn't work well (missing
up-to-date releases.xml repository files, hard to generate obr.xml for
"default" bundles to install, ...).

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>