meter: remove dependency on KillbillTestSuite classes
Now that meter is an osgi bundle, it doesn't have to follow Killbill
core conventions. Let's not bother refactoring its tests to follow the
new Guicy patterns, but get rid at least of the KillbillTestSuite*
dependencies.

Note that tests still depend on the killbill-util test-jar for the
MySQLTestingHelper. I'm not sure what we want to do longer term for
this?

See #38.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>