pom.xml: remove maven-duplicate-finder-plugin from Travis build
It looks like the builds are failing now because the logs are getting
too big:

The command "mvn clean install -Ptravis" exited with 0.

The log length has exceeded the limit of 4 Megabytes (this usually means
that test suite is raising the same exception over and over).

The build has been terminated.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>