jaxrs: revisit the form kludge in PluginResource
It wasn't working in case the Shiro filter was disabled: in this
scenario, nobody was looking at the parameters and for some reason the
inputstream was empty.

This new version is a bit smarter by asking explicitly Jersey for the
form parameters. There are now 2 codepaths for POST requests
(application/x-www-form-urlencoded and others).

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>