payment: revisit refresh call for default payment methods
The refresh call is more complex across plugins - who is authoritative
for the default payment method?

The new strategy is to make Kill Bill solely responsible for the default
payment method and update it only if the plugin associated with the
default payment method is the same as the one we're currently
refreshing.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>