util: fix audit logs retrieval
There were two bugs preventing proper retrieval of audit logs.

First of all, when creating audit logs, the BindBeanFactory will see the TableName
attribute of the EntityAudit object as an enum, and thus it will bind
its name via EnumArgumentFactory (the enum name is written to the database).
When reading audit logs, we were directly binding TableName. The
BindBeanFactory in this case was binding the tableName argument of the enum,
which is the table name string (lower cased): we were comparing
table_name to TABLE_NAME in the queries. This has been fixed: we now
use the enum name for the retrieval queries.

Second, depending on the collation, the queries would return an empty set.
Tests never caught it because the collation wasn't set in our ddl (we
were performing case insensitive comparisons). This has been fixed and closes #101.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>