entitlement: override add-on end date if needed
If a base plan is cancelled prior to an add-on, make sure we return the base plan
cancellation date for the add-on, even if the add-on is effectively cancelled
at a later date (since, if nothing happens, the add-on will be cancelled by the
notification queue when the base plan is cancelled).

See discussion on:

  https://github.com/killbill/killbill-integration-tests/commit/6aa7c6d30699a259983b4c014180fa421d89354a

This fixes test_future_cancel_bp_prior_future_ao_cancel_date in killbill-integration-tests.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>