entitlement: iteration on EventsStreamBuilder speedup optimizations
Explain the magic and add tests for eventsStreamBuilder.buildForAccount. Even though
all tests were passing, I still went ahead and updated buildForAccount to do the same
magic as buildForEntitlement. I didn't manage to write a failing test scenario for it
but it seems required.

This also fixes an NPE with standalone subscriptions.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>