payment: search bugfixes
When listing all payments or payment methods, don't ignore records
that are in Kill Bill but not in the plugins. These exist, regardless
of what the plugin returned.

Also, regarding payments, don't ignore the ones which are associated
with a deleted payment method (the payment method had to be active at
the time of payment).

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>