Fixes #156
There is no easy way to avoid an extra credit on an account that reached a state where blockedBilling is true.

This easiest way to achgieve that result is to have overdue set AUTO_INVOICING_OFF = true when it inserts such a blocking state
and have it remove it when it clears the state: As a result the blocked billing is there in the billing event but invoice will not try to generate
any invoices and when (if) the overdue state clear, then the unblock billing event is inserted and AUTO_INVOICING is removed, which ends up triggereing
the correct credit.