profiles: reduce size of the war
By transitive dependencies, we were including mysql-connector-mxj-db-files
in the war. This is only needed in tests when running an embedded instance
of MySQL. Besides, the jar itself is huge (134M) because it contains versions
for all platforms (FreeBSD, Linux, SunOS, Win) and architectures (i386, x86, sparc).

By removing it, the Kill Bill war is now 45M, down from 178M.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>