subscription: better handle add-on future cancellations
Because of timing issues, add-ons cancelled by the base plan would not always
be reflected in the add-on events list.

This patch revisits the creation of these events by not caring about the time,
but simply looking for any change or cancellation event for the base plan,
which would trigger a cancellation. If the add-on is not yet cancelled,
the earliest event found is now used to add the in-memory cancellation event.

This fixes https://github.com/killbill/killbill/issues/268.

Corresponding integration test: https://github.com/killbill/killbill-integration-tests/commit/d7ec9355d405e384390341c40b38f6e4be827950.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>