catalog: optimize caching
Add a test to verify the default catalog is cached in a multi-tenant
scenario, if no tenant specific catalog has been configured.

The code was already handling the cache miss by adding explicitly an entry
for that tenant (see implementation of EhCacheCatalogCache#getCatalog).
This patch simply ensures that a default catalog is always set.

This fixes https://github.com/killbill/killbill/issues/300.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>