invoice: Refactor the computation of next invoice future notitications
Prior the change, the code would compute those future notifications after having generated the invoice. The code
was cumbersome because some of the info to compute those date were not easily accessible. Also, this does not work
for when we will need to implement the future notification date in the recurring IN_ARREAR because the information would only
be accessible deep inside the invoice computation logic.