invoice: Fix usage bug where we end up missing usage items
The issue has to do with the ordering of the billing events: They are first ordered by subscription and then effectiveDate. The code
assumed that they were ordered by date first and so by using the first event, we had no guarantee to end with the lowest billing event date
(when the account has multiple subscriptions).