invoice, payment: Fixes #415
In addition to fixing the multiple adjustments issue described, the fix also addresses the refund path to make sure that amount is checked in the priorCall from the InvoicePaymentControlPluginApi,
and that the correct exceptionure we only use handling happens. The code change will change the semantics on what happens when we refund +adj after repair .
Also introduces a new invoice index.