util: fix flakiness of TestApiListener#isCompleted
TestConsumableInArrear was sometimes failing due to multiple events being dispatched.

The root cause was the following sequence of events:

* Bus event claimed and dispatched
* Test listener notified
* Other listeners (e.g. invoice) start processing the event
* Clock moved
* Bus re-dispatching the same event (because the claim time has now passed)

The reason the clock is moved too early is because we don't listen to null invoice events in the tests.

We already had a work-around in TestApiListener (we wait until there is no more entries in bus_events), however
one code path was by-passing it. This fixes it -- the main drawback is that we introduce a 500ms delay in the worst
case scenario. We could reduce that arbitrary sleep delay in case it causes a significant slowdown in the tests.

This should fix https://github.com/killbill/killbill/issues/382: I ran the test successfully twice with invocationCount = 100
(while before I could trigger the bug by running the test manually a few times).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>