payment: force overdue refresh from InvoicePaymentControlPluginApi
Due to a race condition, we cannot just rely on payment events to refresh
the overdue state, as the billing state might not be up-to-date yet when
the event is processed.

As a work-around, trigger an overdue refresh after notifying invoice of the payment.

Note that we now take the account lock in overdue, otherwise multiple refreshes would
happen concurrently (the whole overdue process is not atomic, as we modify
entitlements, tags, etc.).

This fixes https://github.com/killbill/killbill/issues/472.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>