subscription: Introduce new EntitlementState.PENDING. See #466
This is the first part of the fix. Currently SubscriptionBase internally on EntitlementState (it should really depend on a SubscriptionBaseState instead as EntitlementState should only be managed
in the entitlement module), so the code inside SubscriptionBase can now be changed to take advantage of that new state and avoid having to handle null cases.