entitlement: allow to pass null dates in APIs
A null date means now: this avoids unnecessary conversions between LocalDate
and DateTime, and gets rid of any logic preventing the creation a future DateTime.

The side effect of this change is that while non-null dates are still respected,
no guarantee is done on when during the day the change will occur (for example,
a subscription cancelled with an effective local date of today might only be effective in
a few hours).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>