junction, invoice: Fixes #481
Modify BillingEvent to include a method 'getRecurringPrice(final DateTime effectiveDate)' and modify FixedAndRecurringInvoiceItemGenerator
to use that method to correctly set the rate for each period invoiced. This allows to correctly find the new recurrring price for catalog
entries which become effective in a later catalog through the use of the effectiveDateForExistingSubscriptons date.