catalog: Remove completely the StandaloneCatalogWithPriceOverride from VersionedCatalog to enable VersionedCatalog XML serialization. See #601
Prior the change VersionedCatalog would contain a list of versions of type StandaloneCatalogWithPriceOverride. StandaloneCatalogWithPriceOverride was implemented
using a delegation pattern to delegate most opeation to the underlysing StandaloneCatalog. The new code will rely on inheritance, which as a result will simplify
StandaloneCatalogWithPriceOverride implemtation and most importantly completely remove StandaloneCatalogWithPriceOverride from VersionedCatalog.

The creators of such VersionedCatalog needs to correctly build the StandaloneCatalogWithPriceOverride for this to work correctly.