subscription, entitlement, jaxrs: Modify entitlement apis to allow operations using either the catalog planName or the undeterministic triplet {productName, billingPeriod, priceList}. See #303.
At the JAXRS level the code still allows operation to happen using the {productName, billingPeriod, priceList} triplet. If the catalog contains multiple entries with the same shape in a given pricelist,
such operations would not be deterministic. Missing is the logic to detect such cases and throw exceptions (along with tests).