subscription: Fix several issues related to future subscription cancellation. See #619.
1. Fix bug in subscription module in the logic where we delete future events when adding a new cancellation event. The logic needs to be
changed to include existing events for the exact same date (which happens when cancelling several times with an EOT policy)
2. Hardening against data integrity issues where we have multiple active CANCEL to skip such invalid events.

As a consequence we may send less events and some tests needed to be slightly modified. Also fixed some broken tests (where we moved the clock too far ahead in one go).