invoice: tighten 3369fdb73b5586ed972f54a1710f478b2001de02
The previous patch was nice as it would prevent any form of double billing,
but it was a bit too aggressive in some cases where we actually do want to re-invoice.

This patch fixes repairs by parts and makes sure we invoice for the full period
in the case of bad repairs for instance (see https://github.com/killbill/killbill/issues/170).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>