invoice, payment: integrate Overdue with PENDING payments
Make invoice send an InvoicePaymentErrorInternalEvent event upon processing
a PENDING payment: this is used as a trigger for Overdue (which doesn't listen
to pure payment events anymore, but just InvoicePaymentInfoInternalEvent / InvoicePaymentErrorInternalEvent).

The name could be a bit misleading but I'm not sure it's worth introducing another
internal InvoicePaymentPendingInternalEvent event for this. The name does match
the state on disk though (success = false while the payment in PENDING).

This internal event will be sent as an INVOICE_PAYMENT_FAILED on the external bus.
Listeners should make sure to retrieve the full payment to act on it (which is
the best practice anyways).

If the payment transitions to success, InvoicePaymentInfoInternalEvent is
subsequently sent (INVOICE_PAYMENT_SUCCESS on the external bus).

See https://github.com/killbill/killbill/issues/625.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>