invoice, payment: second pass at integrating invoice with the AdminPaymentApi
Make sure the link invoice-payment is updated after fixing a transaction
from PAYMENT_FAILURE to SUCCESS.

As part of it, when IncompletePaymentAttemptTask completes a run and the resulting
transaction is success (i.e. SUCCESS or PENDING), the future retry associated
with the completed attempt is removed.

See https://github.com/killbill/killbill/issues/625.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>