cache: switch away from CacheLoader pattern
In Ehcache 3, CacheLoader has been replaced by CacheLoaderWriter. The new interface only
exposes a load(K key) method though, so we are unable to pass additional arguments anymore.

CacheControllerDispatcher, the caching abstraction in Kill Bill bridging to Ehcache,
has now been re-implemented to manually cache values when detecting a miss.

Note that this patch is made while we are still on Ehcache 2 to minimize the changes when
we are ready to upgrade the dependencies (it gives us a known working state to work from).

See https://github.com/killbill/killbill/issues/587.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>