util: make EhCacheBasedCacheController strongly typed
Cache operations on Ehcache 3 require the user to pass the expected
classes for both the key and the value.

To simplify the upgrade, do the same for our cache abstraction layer.

As part of this work, a bug was discovered in DefaultAccountApiBase:
the account id (UUID) was used to populate the ImmutableAccountData cache while
the account record id (Long) was expected instead. As a result, we probably
have a high miss rate on the ImmutableAccountData cache today.

See https://github.com/killbill/killbill/issues/587.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>