account: optimize account creation codepath
During the account creation call, the record isn't yet visible
outside the transaction: any call to ImmutableAccountInternalApi
to refresh the InternalCallContext would not only always
result in a cache miss, but also never result in finding the record
in the database.

Re-work the code to not issue any DB call. This saves us 4 DB queries
per account creation call.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>