util: optimize history handler
The code used to retrieve the entity and the entity record id
both before and after executing the JDBC transaction.

In reality, because we want the latest state for the history tables,
we need to query for the current state before the JDBC transaction
only in case of DELETE operations.

Additionally, no need to retrieve the entity record id separately
(it comes for free from the entity).

This saves 3 JDBC calls for any CREATE or DELETE operation, 2 for any UPDATE operation.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>