invoice: Rework the logic for invoice/account balance. See #749
This change is mostly about writing new test cases for balance computation (DRAFT, WRITTEN_OFF or migrated invoices) and refactoring the code is such a way that it is is easier to understand how to handle such case in each of this specific situations. As a side effect the logic is now more spread throughout the code, but this is precisely because each case may need to have custon logic.