subscription: fix previously uncaught exception in PlanAligner
getNextTimedPhase doesn't find the right plan when processing a PHASE
event when the entry was retired at the time of the PHASE event.

The issue was that the wrong catalog version was pulled: instead of using the
version at the time of the subscription creation, the version at the time of
the phase event was used.

For symmetry, this behavior has been fixed for uncancel operations too (this is
a behavior change as previously uncancel operations would be prevented from the API
when the original plan was retired at the time of the uncancellation).

Additional tests have been written to verify and illustrate these scenarii.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>