util: Modify TagUserApi#getTagDefinitions to include the internal __PARK__ ControlTag.
The __PARK__ ControlTag should only be set in case of data corruption and was previously removed from the TagUserApi#getTagDefinitions because it is an internal tag. However, in scenario where such tag is set, we need
a way for users to resolve the tagDefinitionId using the TagUserApi#getTagDefinitions api. See for instance the
issue reported in the analytics plugin : killbill/killbill-analytics-plugin#69.