subscription: See #795
This fix only takes care of the basic use case where we try to create 2 bundles for the same tenant and  the same external Key. This fix does not handle yet the case of a race condition. We will keep open the issue in `0.20.x` to revisit this specific use case.