subscription, junction: Fix use of date into catalog alignment apis. See #784
Modified code to pass the subscription start date instead of the date of the transition.
The side effect is that if the billing alignment policies were to change across versions
this would not be correctly reflected. The up side, is that deleting a Plan does not throw
an exception anymore.

The real fix is to modify our catalog apis to have the 2 dates, effective date of the change
and subscription start date.