subscription, junction: revisit catalog APIs usage
Avoid fetching the catalog over and over again: when we already have the plan,
we can directly fetch the phase.

This also fixes the implementation of rebuildSubscriptionAndNotifyBusOfEffectiveImmediateChange:
we cannot simply take the old subscription events and add the new event because in some cases, we
have had to de-active some events on disk (e.g. future phase event during a change plan).

It looks like we were incorrectly building subscription events in some cases, even though it didn't
affect most codepaths.

As a side effect, a "spiced-up test" had to be modified (creating garbage data on disk now throws an exception
as the code refuses to build such incorrect timelines).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>