jaxrs: Fix non deterministic behavior in Account#closeAccount by adding AUTO_INVCOICING_OFF tag.
This makes sure that adjusting invoices will only be done on existing unpaid invoices, and will not compete
will potential new generation of invoices due to the cancellation.

For example, a typical scenario is one where we invoiced in advance for a month, moved the clock by one day.
If we let the concurrency happen between subscription events and invoice item adjustments, we may end up with
adjusting before than the invoice subsystem had time to processe the subscription event, and we end up re-invoicing
for a one day.