junction: revisit account BCD calculation
Before, we were using the first calculated BCD of the first BillingEvent seen.
This was a bit arbitrary, especially across bundles (since we weren't looking at the date).

We now look at the first BillingEvent with a non-$0 recurring price and use
its effective date to get the account BCD. This matches better our BCD definition:
date of first recurring non-$0 charge.

One side effect is that for pure usage based plans, the account BCD isn't set anymore.
For instance, a consumable in arrear usage plan billed on the 1st would have previously
set the account BCD to the 1st. Now, it is not updated (unless a recurring price is set to the plan).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>