junction: Introducing a new SubscriptionBillingEvent type to better compute the BillingEvent
This code change also removes a bunch of hacks that we added over the years in the DefaultBillingEvent class.
There are still remaining TODO that will hopefully be tackeled later

Also, we are slowly introducing a new concept of a `lastChangePlanDate` which will be plugged later when fixing the issue
with change of Plan across catalog versions.