invoice: Fix issue in TestInvoiceTrackingSqlDao#testInvalidation.
The test incorrectly made some assertion from audit log targetRecordId to entity recordId -- which was incorrect
since there is a history table in between. In most cases since we only had inserts it worked but if we restart the
tests and record_id gets incremented this would break.

Also, refactor this test to have multiple transactions so we can debug what happens in this complex logic -- as opposed
to having all state in the transaction (incl. assertions and moving the clock) and nothing on disk.