subscription: Change SubscriptionBillingEvent#Plan semantics for cancelled transition. See #1191
Previous code would compute such Plan to be the previous non null Plan which created an asymetric
meaning of what this Plan really mean and how it is compute. Instead we simply return null and need to
update the rest of the code such as junction (BillingEvent) and invoice to correctly understand this.

In addition we expose SubscriptionBillingEvent#getCatalogEffectiveDate which is required (when Plan is null)
and which should provide what we need to implement #1191 (as each BillingEvent now comes with its catalog version).