payment: change processedAmount computation in fixPaymentTransactionState
When going through the AdminPaymentApi, if we don't have plugin details
(e.g. when going through JAX-RS), assume processedAmount is the full amount
in case of success (and 0 in case of failures).

This ensures that the invoice balance will be updated after fixing the state
manually.

See https://github.com/killbill/killbill/issues/1061#issuecomment-521911301

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>