invoice: Filter out $0 REPAIR items.
In scenarios of full item adjustements with a subscription cancelled prior its EOT
the invoice item tree will generate a REPAIR item for the non billable period but because
everything was already credited through full item adj. such item would be generated with a $0 value.

While this makes a lot of sense from a cleanliness of diagrams, this may end up confusing our users, so
filter them out prior writing on disk.