Fix invoice regression introduced in beca83bfe073.
The change introduced in https://github.com/killbill/killbill/commit/beca83bfe073b8fa14a4a248da0a1d1487db84fc#diff-6bbd5e696930ee20c12f24087244c833R234
is actually correct, i.e the recurring billing associated with an event marking cancelation or pause should be set to `NO_BILLING_PERIOD`.

However, our invoicing code contains another bug where we keep adding future notifications for canceled subscriptions. Prior this change,
this was not visible as the notification was set in the future but with the change from commit beca83bfe073, the notification becomes in the past
and what was previously an unused future notification now leads to infinite loops.

The fix is to remove future notifications when there is nothing to invoice in the future.