jaxrs: fix NPE in getPaymentsForInvoice
Handle the case of invoice payments with a NULL payment id (i.e. when
the payment is aborted by a payment control plugin). These won't
be returned by the API.

For testing, I had to add a new (undocumented) query parameter to
the createInstantPayment API, in order to add an additional payment
control plugin to the chain.

This fixes https://github.com/killbill/killbill/issues/1288.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>