Fix issue when using AUTO_INVOICING_REUSE_DRAFT and with (subscription) dates in the past.
The scenario is one where we have AUTO_INVOICING_DRAFT/AUTO_INVOICING_REUSE_DRAFT set on the account
and we create some subscriptions in the past.

There will be an initial bus event for the start date creating the invoice and adding the first RECURRING period (from test example: 8/7 -> 9/1)
Then, a notification be picked up immediately for the next RECURRING period as its effective date is in the past (9/1 -> 10/1)

Unfortunately prior the fix the targetDate which is computed as the max(event date, all invoices targetDate) is wrong because targetDate from (resused) invoice
was not updated.

This fix will now update the targetDate associated with the ongoing (reused) invoice and therefore avoid all the REPAIR we were seeing.