jaxrs: Update AccountResource#payAllInvoices to return a Location header indicating paid invoices
This is an api change which modifies the behavior of the AccountResource#payAllInvoices to return a 201 + Location header instead of the 204 previously.

In addition, we also implemented a new filter 'invoicesFilter' to the endpoint AccountResource#getInvoicesForAccount to allow the follow up of the request, allowing clients to automatically fetch the resulting paid invoices.