catalog: Modify key for internal plan override cache
Prior this change, we would use the `planName` as the key for the plan override cache.
This issue is that we would return the non overriden Plan entry regardless of the catalog version.

Therefore, when using the catalog api `findCatalogPlanEntry(..., requestedDate, subscriptionChangePlanDate)`, we would not
stop at the entry where such Plan is defined but instead go through all versions whose effective dates > requestedDate
(even if the non overriden Plan does not exist in such version).

After this change, we concatenate the catalog effective date (using the `!` delimiter) to solve the issue.