invoice: revisit account BCD update event handling
Upon receiving a BCD update event, simply trigger an invoice run.
This is more in line with how the system would behave upon processing
subscription BCD change effective events. In practice, this means repairs
are generated right away, instead of waiting for the next BCD.

This uncovered a small issue with how next notifications are computed:
specifically, when only a REPAIR_ADJ was generated, we were clearing
the notifications entirely. As a side effect, we were not generating
the invoice at the next BCD (2016-7-17) in testAccountBCDChangeWithNoOptimization.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>