invoice: Complete rewrite of NodeInterval#addNode
Prior this fix, we had an implementation of NodeInterval#addNode that was left with lots of preconditions for cases not implemented,
and the very recent addition in c6c569d98d09730 and 07eb86f3fbf2f clearly show we are not on the right track.

After carefull analysis, there is a much better and simpler way to add node intervals into the tree (code now supports any use cases and is much simpler).

However, there are some behavior changes on how this tree is now constructed since we don't 'rebalance' nodes and end up with more splits.

The code is functionally equivalent for all use cases (all beatrix tests pass as-is), but specific invoice tests require some changes.