invoice: Modify and add tests for change of behavior introduced in 6338109f8cdc7f21
The re-implemnentation of the NodeInterval#addNode changes a behavior for a case where there was a full invoice item adjustment associated with an invalid recurring item (a very edge case). I wrote a beatrix test to reproduce the scenario with the current implemention prior this fix and realized it lead to a 'double billing detected' exception.

With the new implementation we end up on a stable state and with something that is very sane (e.g we re-invoice for the small piece missing from the original wrong item).