invoice: Fix TestDefaultInvoiceGenerator#testRegressionFor170 with new behavior. See 6338109f8cdc7f21
With the new construction of the tree, Kill Bill generates a new 'equivalent' behavior so the expectations on the test TestDefaultInvoiceGenerator#testRegressionFor170 need to be changed.

In order to verify that old generated data for such scenario will not cause additional invoicing with new behavior, I created a beatrix level
test to reproduce the scenario along with the old data and ensure that invoice generation leads to a no-op.