invoice: Fix a ILLEGAL INVOICING STATE in repair + adj use case. See #1505
It seems like we had some logic long time ago to remove cancelling pair of items on the same interval (i.e a RECURRING item fully REPAIRED) but this was removed in `https://github.com/killbill/killbill/commit/789ed9ff15` and yet the comments and logic in the code would still refer to this mechanism `https://github.com/killbill/killbill/blob/killbill-0.22.27/invoice/src/main/java/org/killbill/billing/invoice/tree/ItemsInterval.java#L178`.

For most cases where we have a full repair, we rely on the `InvoicePruner` to remove both the RECURRING and its fully REPAIRED item prior we even entre the tree, as shown in `https://github.com/killbill/killbill/blob/killbill-0.22.27/invoice/src/main/java/org/killbill/billing/invoice/generator/InvoicePruner.java#L206`.

However, the scenario we have here is one where we have a fully adjusted RECURRING item, and then an additional $0 `REPAIR_ADJ` item and in such a scenario,
the `InvoicePruner` will keep both the `ITEM_ADJ` and the `REPAIR_ADJ`, leading to a case that is not handled once we are in the tree - or rather a case that triggers a precondition.
Removing the precondition would lead to keep generating more $0 `REPAIR_ADJ` on every run, not a good thing...

The initial attempt to fix this issue was to extend the code in `InvoicePruner` and it can be made to work, but it slightly changes the behavior for an edge case scenario where the system tries to be smart about incorrect `RECURRING` items - see TestFixedAndRecurringInvoiceItemGenerator#testItemPartiallyRepairedAndPartiallyAdjusted. It is debatable whether we really care about keeping the behavior we have, but in order to minimize disruption, I gave up on this implementation.

Instead, I added some logic to harden the cases where we have canceling pairs of RECURRING item fully REPAIRED inside the interval tree to support this use case.

I had to change a a unit test that would also expect a case where we have unexpected RECURRING items on disk to fail, to instead return 0 items (ignore it).

إِنْ شَاءَ ٱللَّٰهُ