payment: Introduce a new property to allow updating a payment method
By default, we only allow *setting* a new payment method from a payment control plugin.

When this new property is set to true, we allow overwriting whatever value it already had.

This needs to be used with caution, as we could have multiple updates from multiple plugins leading to a non deterministic state, or we could attempt to update a payment method for an existing `Payment` and it is not clear what the end state would be. See #1097