subscription: Provide a new api `SubscriptionBaseInternalApi#updateChargedThroughDates`
The old api `SubscriptionBaseInternalApi#updateChargedThroughDate` has been removed and we now allow to update many ctds in one call.

We also simplified the logic and eliminated the use of updating `bundle#getLastSysUpdateDate`. See d4d12979a1

Note that the impl. relies on a sql `IN` clause which could become an issue with very large dataset and depending on the engine.