subscription, jaxrs: Provide a (hacky) way to specify Subscription keys during transfer. See #1627
Beacuse of apis restrictions, we rely on plugin properties to pass this information.

The format of the (plugin properties) keys is undocumented and is subject to change in future releases; it is namespaced with `KB_` to minimize colisions
with possible use cases of entitlement plugin using their own keys.

Proper fix will be introduced in `0.24` release.