subscription: Fix issue in bundle transfer
In cases where we have a BCD_UPADTE event, the code would throw an `java.lang.IndexOutOfBoundsException` because the event was not correctly handled.
This fix makes sure that when we have both a CREATE and BCD_UPADTE prior the transfer date, the code does not throw exception and also transfers the BCD alignment.