subscription: Modify subscription events sql interface to provide ordering
The methods from `SubscriptionEventSqlDao` now return a `SortedSet` which leverages
the `Comparator` implementation introduced in c76715b1728f5. As a result we don't need
to explicitely call `Collections.sort` in our dao layer.

However, we cannot use the shared `EntitySqlDao#getByAccountRecordId` method anymore so I refactored the code to not use such method.

Also added a test to verify the ordering is correctly implemented.