subscription: Add bundle transfer test
We verify the behabior of doing a transfer in the past prior the start date of the original subscription.

Production use case had shown some bad data being created but it looks like the fix introduced in 0a579dde34 solves the issue
by realigning the dates to the start of the subscription.