subscription: Fix issue with computation of local BCD when aligned with other transition (e.g CREATE)
Our billing events return the local BCD if applicable, but the computation was not precise when we have a CREATE and BCD_CHANGE on the same date,
as we would only discover the value when parsing the next transition. The code now makes sure that for each transition, we populate the local BCD with the right value.

Note: This would not directly impact the invoicing of such use case but this showed up in other scenarios, such as highlighted in the new test `TestCatalogWithEffectiveDateForExistingSubscriptionsCustomConfig#testSubscriptionNotAlignedWithVersionChange2`