beatrix: Harden usage tests
With the changes introduced in https://github.com/killbill/killbill/pull/1738/commits/08f9b913b886b8e497b6edde02d51f4166137e69, we need to be more cautious
when we record usage on the first day, to make sure the timestamp is not too early and usage data points is indeed taken into consideration.

For such test scenario we can use the test api `recordUsageData` which uses a DateTime and align with the clock value.

I did a pass on all usage tests, but this may be part of an incremental effort to fully stabilize this.