beatrix: Change expected behavior in TestCatalogWithEffectiveDateForExistingSubscriptionsCustomConfig#testSubscriptionNotAlignedWithVersionChange4
In 2fba9e04b5b, we updated the behavior of how we create the billing event for the catalog effective date when the property  `org.killbill.subscription.align.effectiveDateForExistingSubscriptions=true`.

As a result, we need to update this existing beatrix test. I actually think it makes more sense to way we have it now as we don't apply a new price
until we explicitly pass the Subscription#effectiveDateForExistingSubscriptions date. But it's really a matter of semantics.