invoice: Fix usage issue for non UTC Account
The computation of the `ContiguousIntervalUsageInArrear#toEndOfDay` was not correctly taking into account
the Account#TZ. As a result, we end up creating endOfDay `transition` which are not the end of day as seen from
the customer Account.

When later in the method `ContiguousIntervalUsageInArrear#computeMissingItemsAndNextNotificationDate` we compute a LocalDate based
on this endOfDay `transition`, we do not end up with the correct transition; during the next invoice run, the method
`ContiguousIntervalUsageInArrear#getBilledItems` which should return the existing USAGE invoice item(s) for a given period
does not find an existing item leading to a double billing of the period.