subscription: Fix billing event logic for catalog transition with multi-phase plans
The tests TestCatalogWithEffectiveDateForExistingSubscriptionsCustomConfigMultiPhasePlan
showed a dupplicate transition was being added in our billing events for plans with multiple phases
and multiple versions with `effectiveDateForExistingSubscriptions`.

We need to check if the catalog transition was already added as part of a previous plan transition.

This undo the fix introduced in f58cba4edc31ff5