subscription: Fix issue with port from 0.24. See #1862
The port from 0.24 introduced in 5803afb14ceb5ca4be removed the existing variable `lastPhaseTransition`
which was not used in 0.24, but which was used in 0.22 because of the different handling of the `EXPIRED` event.

This fix reintroduces the variable and adds the new one `lastPlanTransition` (both of them are needed in 0.22).