invoice: Fix issue in usage with daylight saving - See #1934
We introduce a different logic for usage code to convert dates to guarantee some stability across accounts created with the same TZ but at different dates during the year (daylight saving or standard). We isolated the code in such a way that we could easily revert
(or make it configurable) if needed.