invoice: Fix issue when fetching usage points for account in non UTC timezone. See #1981
The `RawUsageOptimizer#getInArrearUsage` method takes a targetDate (`LocalDate`) in the account#tz and it needs to
decide which points to pull, so it needs to be converted in a `DateTime`. We can leverage the `UsageClockUtil#toDateTimeAtEndOfDay`
to get a date at the end of the day with respect to the `account#tz`, and guarantee we don't miss any points.