catalog: Rewrite implementation of CatalogOverrideTierBlockSqlDao#getTargetTierDefinition
Previous implementation relied on sql `concat` method which seem to prevent sql engine from using an index.

The new implementation simplifies the logic by looking at the first block only (with the assumption) and all blocks
within a tier should provide the matching tier (if this were not the case previous impl would anyway fail to return the entry).