subscriptions: Modify billing events for overriden plans.
For overriden plans, we return a `Plan` object that points to a catalog, but the
effective date of this Plan should match the date at which the event was effective and not the original
catalog date.

Failure to do so, means that we don't compute the correct usage intervals: Such intervals are created
when we discover a new transition identified with a key made of `UsageKey={usageName, catalogVersion}`.

Also, remove fix in `SubscriptionUsageInArrear#computeInArrearUsageInterval` introduced in 24543fea7