Modify impl. of `DefaultSubscriptionBase#getSubscriptionBillingEvents` to correctly handle overriden plans. See #2066
In 7d975d69, we introduced a a fix for #2066 but the impl. required going through every plan entries for each billing event (which for large catalog would be a problem).
Instead, we now expose a core functionality from the catalog module to detect overriden plans (based on naming convention). This requires creating a new interface `PriceOverrideSvcStatus`
which is passed around from catalog module to subscription module.

Note: A better way to do it would be to mofiy our Plan interface to expose a method `isOverriddenPlan()`, but that would require changting our APIs.