invoice: Fix wrongly aligned date for cases of end-of-month BCD. See #2133.
The method InvoiceDateUtils#advanceByNPeriods simply adds a number of fixed 'periods' but does not try to align with a possible end-of-month BCD.

In most cases, BillingIntervalDetail already follows the use of this method with `BillCycleDayCalculator#alignProposedBillCycleDate` but not always,
and this is the cause of the issue we are seeing in #2133.

Also note, there is already a method `BillCycleDayCalculator#getFutureBillingDateFor` that calls bnopth methods. So the fix is to always call this method.